% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_validation.R
\name{cross_validation}
\alias{cross_validation}
\title{Fonction cross_validation}
\usage{
cross_validation(X, Y, ncomp)
}
\arguments{
\item{X}{(The X)}

\item{Y}{(The Y)}

\item{ncomp}{(The number of component to be used)}
}
\description{
Fonction cross_validation
}
\keyword{fit,}
\keyword{pls,}
\keyword{plsda}
