% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit}
\alias{fit}
\title{Fonction fit}
\usage{
fit(par = "PLSDA")
}
\arguments{
\item{par}{(par defaut, "toi")}
}
\description{
Fonction fit
}
\examples{
fit("test")
}
\keyword{fit,}
\keyword{pls,}
\keyword{plsda}
