% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.plsda.R
\name{plot.plsda}
\alias{plot.plsda}
\title{Fonction plot.plsda}
\usage{
\method{plot}{plsda}(x, ...)
}
\arguments{
\item{x}{A plsda object containing a trained model. The model can be generated with the fit function of this package.}

\item{...}{Use to comply generic model}
}
\value{
None.
}
\description{
Plot the quality of the model and the VIP for each variable and component.
}
\examples{
model <- fit(Species ~ ., data = iris)
plot(model)
}
\keyword{VIP,}
\keyword{pls,}
\keyword{plsda}
\keyword{quality,}
