% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pls.R
\name{pls}
\alias{pls}
\title{Fonction pls}
\usage{
pls(X, Y, nc, cv, nfold)
}
\arguments{
\item{X}{The predictors must be a set of quantitative variables.}

\item{Y}{The response must be a qualitative variable.}

\item{nc}{The number of component to compute. If cross-validation is activated it will be used as the maximum number of component to return.}

\item{cv}{If TRUE a cross-validation will be done. Default is FALSE.}

\item{nfold}{If cv parameter is TRUE, nfold will specify the kind of cross-validation. Default is 0.
A value of 0 will be equivalent to a "Leave One Out" cross-validation. If > 0 it will do a k-fold cross-validation with k = nfold.}
}
\value{
A plsda object that contains the component matrix, the trained model as coefficients, the explained variance on X and Y, VIP values, quality of the model and cross-validation results if asked
}
\description{
Generate a plsda model according to formula and data provided. This model can be cross-validated.
}
\keyword{fit,}
\keyword{pls,}
\keyword{plsda}
