% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.plsda.R
\name{predict.plsda}
\alias{predict.plsda}
\title{Fonction predict}
\usage{
\method{predict}{plsda}(model, X, y.exp = NULL)
}
\arguments{
\item{model}{(The model to use)}

\item{newdata}{(The test data)}
}
\description{
Fonction predict
}
\examples{
predict()
}
\keyword{fit,}
\keyword{pls,}
\keyword{plsda}
