% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict}
\alias{predict}
\title{Fonction predict}
\usage{
predict(model, newdata)
}
\arguments{
\item{model}{(The model to use)}

\item{newdata}{(The test data)}
}
\description{
Fonction predict
}
\examples{
predict()
}
\keyword{fit,}
\keyword{pls,}
\keyword{plsda}
