% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.plsda.R
\name{summary.plsda}
\alias{summary.plsda}
\title{Fonction summary.plsda}
\usage{
\method{summary}{plsda}(object, ...)
}
\arguments{
\item{object}{A plsda object containing a trained model. The model can be generated with the fit function of this package.}

\item{...}{Use to comply generic model}
}
\value{
None.
}
\description{
Displays a plsda object attributes.
}
\examples{
model <- fit(Species ~ ., data = iris)
summary(model)
}
\keyword{VIP,}
\keyword{pls,}
\keyword{plsda}
\keyword{predict,}
\keyword{quality,}
\keyword{summary,}
