% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validation}
\alias{validation}
\title{Fonction validation}
\usage{
validation(result, y.exp)
}
\arguments{
\item{result}{The result obtained with the predict.plsda of this package, must contains y.hat and Y.hat as a list.}

\item{y.exp}{The expected y, it will be confronted to y.hat found in the result parameter, so it must have the same number of items/}
}
\value{
A plsda-valid object that contains residuals, confusion matrix and error rate.
}
\description{
This function takes a plsda-pred, a prediction computed by the predict function of the plsda package and compare it to the y.exp vector provide.
    It will returns the residuals, the confusion matrix and the error rate.
}
\examples{
model <- fit(Species ~ ., data = iris)
result <- predict(model, iris[,1:4])
validation(result, iris[,5])
}
\keyword{confusion}
\keyword{matrix,}
\keyword{pls,}
\keyword{plsda}
\keyword{predict,}
\keyword{residuals,}
