% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_select.R
\name{var_select}
\alias{var_select}
\title{Fonction var_select}
\usage{
var_select(model, vs_crit, X, Y, ncomp)
}
\arguments{
\item{model}{A plsda object containing a trained model. The model can be generated with the fit function of this package.}

\item{vs_crit}{Treshold value for the variables selection (vs=TRUE). Default value : 1 (fit fonction)}

\item{X}{The X matrix of predictors}

\item{Y}{The Y matrix of binarized response}

\item{ncomp}{The number of component to be used}
}
\description{
This function takes a plsda model and selects variables from the X matrix using the VIP values returned.
}
\keyword{VIP,}
\keyword{fit,}
\keyword{pls,}
\keyword{plsda,}
\keyword{selection}
\keyword{variable,}
