% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit}
\alias{fit}
\title{Fonction fit}
\usage{
fit(formula, data, ncomp = 2, cv = FALSE, nfold = 0, vs = FALSE,
  vs_crit = 1)
}
\arguments{
\item{formula}{A valid statistical R formula.}

\item{data}{The data to be used to rain the PLS-DA model. I must fit to the formula.
The response must be a qualitative variable. The predictors must be a set of quantitative variables.}

\item{ncomp}{The number of component to compute. If cross-validation is activated it will be used as the maximum number of component to return.}

\item{cv}{If TRUE a cross-validation will be done. Default is FALSE.}

\item{nfold}{If cv parameter is TRUE, nfold will specify the kind of cross-validation. Default is 0.
A value of 0 will be equivalent to a "Leave One Out" cross-validation. If > 0 it will do a k-fold cross-validation with k = nfold.}

\item{vs}{If TRUE an automatic variable selection will be done. Default is FALSE.}

\item{vs_crit}{Treshold value for the variables selection (vs=TRUE). Default value : 1 (fit fonction)}
}
\value{
A plsda object that contains the component matrix, the trained model as coefficients, the explained variance on X and Y, VIP values, quality of the model and cross-validation results if asked
}
\description{
Generate a plsda model according to formula and data provided. This model can be cross-validated.
}
\examples{
fit(Species ~ ., data = iris)
fit(Species ~ ., data = iris, 2)
fit(Species ~ ., data = iris, vs = TRUE)
fit(Species ~ ., data = iris, 4, cv = TRUE, nfold = 10)
}
\keyword{fit,}
\keyword{pls,}
\keyword{plsda}
